Delete Scrabble Game Reflection.docx
I'm deleting my file because I am submitting an updated reflection statement. I updated it because it mentions that one of the problems encountered was not receiving a response from my classmates, which I literally just found out a couple minutes ago was not true. It was all my computer, it was acting weird. 
I do not want to discredit them. I hope this explanation helps.